evaluation/style_loss_metric/requirements.txt
numpy 
Pillow
torch
torchvision
tqdm
